backend: Re-write HTTP server with gotham.
Warp has a unique and interesting API, but understanding the type errors
was insane. It's a huge mess of closed types, and it was impossible to
understand what was going wrong, even when doing simple stuff like
trying to change routes based on server config.

Gotham was *much* easier to get going with, and I can kinda understand
the API.